==> notebooks/1-data-collection.ipynb <==
# Data Collection

To collect all the needed training data, I need to use mostly of the same methods I used to get the sample data. But instead of getting all the reviews from one game, I need a wide variety of games. I start by getting every App ID on Steam, over 115,000. I then grab the first 50 reviews labeled "most helpful". Many games had no reviews, or less than 50, especially when only taking english reviews into account. I end up with over one million reviews for this dataset.

## Getting App IDs

import pandas as pd
import requests

response = requests.get('https://api.steampowered.com/ISteamApps/GetAppList/v2/')
app_ids_df = pd.DataFrame(response.json()['applist']['apps'])
app_ids_df.to_pickle('../data/app_ids.pkl.gz')
app_ids_df

app_ids = app_ids_df['appid'].to_numpy()
app_ids, len(app_ids)

## Getting Reviews

def get_reviews(appid, params):
        url_start = 'https://store.steampowered.com/appreviews/'
        try:
            response = requests.get(url=url_start+str(appid), params=params, headers={'User-Agent': 'Mozilla/5.0'})
        except:
                return {'reviews' : []}
        return response.json() # return data extracted from the json response

reviews = []
cursor = '*'
params = { # https://partner.steamgames.com/doc/store/getreviews
    'json' : 1,
    'filter' : 'all', # sort by: recent, updated, all (helpfullness)
    'language' : 'english', # https://partner.steamgames.com/doc/store/localization
    'day_range' : 9223372036854775807, # shows reveiws from all time
    'review_type' : 'all', # all, positive, negative
    'purchase_type' : 'all', # all, non_steam_purchase, steam
    'num_per_page' : 50,
    'cursor' : '*'.encode()
          }

for i, app_id in enumerate(app_ids):
    reviews += get_reviews(app_id, params)['reviews']
    if (i+1)%500 == 0:
        print(f'{i+1} of {len(app_ids)}: {len(reviews)} reviews')

When running the above code to get the reviews, I got the following error:

JSONDecodeError: Unexpected UTF-8 BOM (decode using utf-8-sig): line 1 column 1 (char 0)

It seems to be fixable by changing the encoding in the requests module, but seeing as this code takes about half a day to run, and I have over a million reviews anyways, this will be good enough.

len(reviews)

reviews_df = pd.DataFrame(reviews)[['review', 'voted_up']]
reviews_df.dropna(inplace=True)
reviews_df.reset_index(inplace=True)
reviews_df

reviews_df.voted_up.value_counts(normalize=True)

reviews_df.to_pickle('../data/reviews_raw.pkl.gz')

The initial dataset for this project used games taken from the "hot" section of Steam, and had a 80-20 class imbalance. This data, taken from all Steam apps, has a 75-25 class imbalance. Popular games are more likely to have more reviews, and also have more DLC, which counts as its own app for review purposes. This imbalance might be improved by only counting games, and by taking less reviews per game. However, that would end up with significantly less data overall, so I have decided not to go there for now.

As well, a future improvement could be to get data from other sources in addition to Steam. Metacritic seems like a good choice, as its reviews come with scores. There are also storefronts, such as itch.io or GOG that cater to different types of games than Steam or Metacritic, and so might help the model's ability to generalize.

However, at present, adding more data would just prevent my computer from running these models at all. If I want to increase the dataset size, I first need to get these notebooks running on a better computer, or on something like Amazon Sagemaker.

## Splitting and Pickling

The raw reviews data is too large for my computer to handle, so here I split it into 10 parts and save them seperately. I also randomize the dataframe before splitting.

import numpy as np

reviews_df = pd.read_pickle('../data/reviews_raw.pkl.gz')
reviews_df = reviews_df.sample(frac=1).reset_index(drop=True)

for i, df in enumerate(np.array_split(reviews_df, 10)):
    df.to_pickle(f'../data/reviews_raw_{str(i)}.pkl.gz')

The below functions are ones I had used in old versions of gathering the data. They are no longer used.

def get_n_reviews(appid, n=100):
    reviews = []
    cursor = '*'
    params = { # https://partner.steamgames.com/doc/store/getreviews
            'json' : 1,
            'filter' : 'all', # sort by: recent, updated, all (helpfullness)
            'language' : 'english', # https://partner.steamgames.com/doc/store/localization
            'day_range' : 9223372036854775807, # shows reveiws from all time
            'review_type' : 'all', # all, positive, negative
            'purchase_type' : 'all', # all, non_steam_purchase, steam
        }
    while n > 0:
        params['cursor'] = cursor.encode() # for pagination
        params['num_per_page'] = min(100, n) # 100 is the max possible reviews in one requests
        n -= 100
        
        try:
            response = get_reviews(appid, params)
        except:
                return []
        
        cursor = response['cursor']
        reviews += response['reviews']
        
        if len(response['reviews']) < 100: break
    
    return reviews

from bs4 import BeautifulSoup

def get_n_appids(n=100, filter_by='topsellers'):
    appids = []
    url = f'https://store.steampowered.com/search/?category1=998&filter={filter_by}&page='
    page = 0
    
    while page*25 < n:
        page += 1
        response = requests.get(url=url+str(page), headers={'User-Agent': 'Mozilla/5.0'})
        soup = BeautifulSoup(response.text, 'html.parser')
        for row in soup.find_all(class_='search_result_row'):
            appids.append(row['data-ds-appid'])
    
    return appids[:n]
