==> notebooks/4-base-models.ipynb <==
# Base Models

This processed data is not uploaded to the Github repo, as some of the files are too large. Run notebook 2 in order to produce the same files.

## Baseline Models

Even though I ran this type of model analysis with the sample data, I want to run it again now that I've created bigrams. I'm also no longer using SVM, as it took way too long to run even on the smaller dataset. Once again, logistic regression is the best performing model. Also as expected, bigrams improved the accuracy of the models.

import pandas as pd
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
from sklearn.dummy import DummyClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB
from sklearn.ensemble import RandomForestClassifier

def get_model_metrics(X_train, y_train, X_test, y_test, model, model_name, data_name):
    
    model.fit(X_train, y_train)
    y_train_hat = model.predict(X_train)
    y_test_hat = model.predict(X_test)
    
    acc_train = accuracy_score(y_train, y_train_hat)
    pre_train = precision_score(y_train, y_train_hat)
    rec_train = recall_score(y_train, y_train_hat)
    f1_train = f1_score(y_train, y_train_hat, average='macro')
    
    acc_test = accuracy_score(y_test, y_test_hat)
    pre_test = precision_score(y_test, y_test_hat)
    rec_test = recall_score(y_test, y_test_hat)
    f1_test = f1_score(y_test, y_test_hat, average='macro')
    
    metrics = {'Model': model_name,
               'Processing': data_name,
               'Test Accuracy': acc_test,
               'Test Precision': pre_test,
               'Test Recall': rec_test,
               'Test F1': f1_test,
               'Train Accuracy': acc_train,
               'Train Precision': pre_train,
               'Train Recall': rec_train,
               'Train F1': f1_train}
    
    return metrics

"""
datasets = [('TF-IDF', 'tf'),
        ('TF-IDF with Bigrams', 'bigram'),
        ('Document Embeddings', 'embed')]
"""
datasets = [('TF-IDF with Bigrams', 'bigram')]
models = [('Logistic Regression', LogisticRegression(solver='saga')),
          ('Multinomial Naive Bayes', MultinomialNB()),
          ('Random Forest', RandomForestClassifier())]
metrics = []

y_train = pd.read_pickle('../data/processed/y_train.pkl.gz')['voted_up'].to_numpy()
y_test = pd.read_pickle('../data/processed/y_test.pkl.gz')['voted_up'].to_numpy()

for data_name, file in datasets:
    X_train = pd.read_pickle(f'../data/processed/X_{file}_train.pkl.gz').to_numpy()
    X_test = pd.read_pickle(f'../data/processed/X_{file}_test.pkl.gz').to_numpy()
    for model_name, model in models:
        print(model_name, data_name)
        metrics.append(get_model_metrics(X_train, y_train, X_test, y_test, model, model_name, data_name))

metrics.append(get_model_metrics(X_train, y_train, X_test, y_test, DummyClassifier()))

metrics_df = pd.DataFrame(metrics)
metrics_df.sort_values(by='Test Accuracy', ascending=False)

# Gridsearch

Here I performed a gridsearch on the random forest and logistic regression models using just the bigram data, as it performed the best. Naive Bayes models do not have any hyperparameters to tune, and so there is no grid search to perform on it.

import pandas as pd
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import GridSearchCV

y_train = pd.read_pickle('../data/processed/y_train.pkl.gz')['voted_up'].to_numpy()
X_train = pd.read_pickle('../data/processed/X_bigram_train.pkl.gz').to_numpy()

param_grid_lr = {'C': [0.1, 1, 10],
                 'class_weight': ['balanced', None],
                 'solver': ['saga']}
gs_lr = GridSearchCV(estimator=LogisticRegression(), param_grid=param_grid_lr, scoring='f1_macro', cv=3, verbose=5)
gs_lr.fit(X_train, y_train)
gs_lr.best_params_

param_grid_rf = {'n_estimators': [100, 500],
                 'max_features': ['auto', 150],
                 'class_weight': ['balanced', None]}
gs_rf = GridSearchCV(estimator=RandomForestClassifier(), param_grid=param_grid_rf, scoring='f1_macro', cv=3, verbose=5)
gs_rf.fit(X_train, y_train)
gs_rf.best_params_

## Final Models

After comparing the best tuned models, logistic regression still has the best accuracy. I had expected random forest to improve performance more with tuning, but it seems not to be the case here.

y_train = pd.read_pickle('../data/processed/y_train.pkl.gz')['voted_up'].to_numpy()
y_test = pd.read_pickle('../data/processed/y_test.pkl.gz')['voted_up'].to_numpy()
X_train = pd.read_pickle('../data/processed/X_bigram_train.pkl.gz').to_numpy()
X_test = pd.read_pickle('../data/processed/X_bigram_test.pkl.gz').to_numpy()

lr_final = LogisticRegression(C=10, solver='saga')
metrics = get_model_metrics(X_train, y_train, X_test, y_test, lr_final, 'Logistic Regression', 'TF-IDF with Bigrams')
print(metrics)

metrics

lr_final = LogisticRegression(C=10, solver='saga')
nb_final = MultinomialNB()
rf_final = RandomForestClassifier(max_features=150)

final_metrics = []
print('starting Logistic Regression model')
final_metrics.append(get_model_metrics(X_train, y_train, X_test, y_test, lr_final, 'Logistic Regression', 'TF-IDF with Bigrams'))
print('starting Naive Bayes model')
final_metrics.append(get_model_metrics(X_train, y_train, X_test, y_test, nb_final, 'Multinomial Naive Bayes', 'TF-IDF with Bigrams'))
print('starting Random Forest model')
final_metrics.append(get_model_metrics(X_train, y_train, X_test, y_test, rf_final, 'Random Forest', 'TF-IDF with Bigrams'))
print('completed models')

final_metrics_df = pd.DataFrame(final_metrics)
final_metrics_df.sort_values(by='Test Accuracy', ascending=False)

## Save Model

Even though I will also be creating a neural network model, I still want to save this best logistic regression model. I can try to use it as a backup in case the neural network model is too big to upload to heroku.

import pandas as pd
from sklearn.linear_model import LogisticRegression
import pickle

y_train = pd.read_pickle('../data/processed/y_train.pkl.gz')['voted_up'].to_numpy()
y_test = pd.read_pickle('../data/processed/y_test.pkl.gz')['voted_up'].to_numpy()
X_train = pd.read_pickle('../data/processed/X_bigram_train.pkl.gz').to_numpy()
X_test = pd.read_pickle('../data/processed/X_bigram_test.pkl.gz').to_numpy()

model = LogisticRegression(C=10, solver='saga')
model.fit(X_train, y_train)

pickle.dump(model, open('../final_model/sklearn-logreg/model.pk', 'wb'))
